+++ admin/src/scenes/plan-transport/schema-repartition/group/GroupAffectationSummary.js
@@ -6,8 +6,11 @@ import People from "../../../../assets/icons/People";
 import api from "../../../../services/api";
 import IcebergColor from "../../../../assets/icons/IcebergColor";
 import MapColor from "../../../../assets/icons/MapColor";
+import { useSelector } from "react-redux";
+import { ROLES } from "snu-lib";
 
 export default function GroupAffectationSummary({ group, className = "", onChange, onChangeStep }) {
+  const { user } = useSelector((state) => state.Auth);
   const [loading, setLoading] = useState(true);
   const [detail, setDetail] = useState(null);
   const [error, setError] = useState(null);
@@ -28,11 +31,11 @@ export default function GroupAffectationSummary({ group, className = "", onChang
       }
       setLoading(false);
     })();
-  }, []);
+  }, [group]);
 
   return (
     <GroupBox className={className}>
-      <GroupHeader onBack={() => onChangeStep(GROUPSTEPS.MODIFICATION)}>
+      <GroupHeader onBack={() => onChangeStep(GROUPSTEPS.MODIFICATION)} noBack={user.role === ROLES.REFERENT_DEPARTMENT}>
         <div className="grow flex items-center justify-between">
           <div className="">Récapitulatif</div>
           <div className="flex items-center text-base text-gray-900">
@@ -54,7 +57,13 @@ export default function GroupAffectationSummary({ group, className = "", onChang
               <div className="grow">
                 <div className="text-lg font-bold text-[#242526]">Centre de cohésion</div>
                 <div className="text-sm text-gray-800">
-                  {detail.center.name}, {detail.center.address} {detail.center.zip} {detail.center.city}&nbsp;•&nbsp;{detail.center.department}
+                  {detail.center ? (
+                    <span>
+                      {detail.center.name}, {detail.center.address} {detail.center.zip} {detail.center.city}&nbsp;•&nbsp;{detail.center.department}
+                    </span>
+                  ) : (
+                    <span className="opacity-50">Pas de centre sélectionné.</span>
+                  )}
                 </div>
               </div>
             </div>
@@ -62,24 +71,30 @@ export default function GroupAffectationSummary({ group, className = "", onChang
               <MapColor className="w-[50px] h-[50px] mr-4" />
               <div className="grow">
                 <div className="text-lg font-bold text-[#242526]">Lieux de rassemblement</div>
-                <ul className="list-outside ml-3">
-                  {detail.gatheringPlaces.map((pdr) => (
-                    <li key={pdr._id} className="text-sm text-gray-800">
-                      {pdr.name}, {pdr.address} {pdr.zip} {pdr.city}&nbsp;•&nbsp;{pdr.department}
-                    </li>
-                  ))}
-                </ul>
+                {detail.gatheringPlaces && detail.gatheringPlaces.length > 0 ? (
+                  <ul className="list-outside ml-3">
+                    {detail.gatheringPlaces.map((pdr) => (
+                      <li key={pdr._id} className="text-sm text-gray-800">
+                        {pdr.name}, {pdr.address} {pdr.zip} {pdr.city}&nbsp;•&nbsp;{pdr.department}
+                      </li>
+                    ))}
+                  </ul>
+                ) : (
+                  <span className="opacity-50">Aucun point de rassemblement sélectionné.</span>
+                )}
               </div>
             </div>
           </>
         )}
       </div>
-      <div className="flex items-end justify-center pt-[67px]">
-        <BorderButton onClick={() => onChangeStep(GROUPSTEPS.MODIFICATION)} className="mr-[8px]">
-          Annuler
-        </BorderButton>
-        <PlainButton onClick={() => onChange(group, GROUPSTEPS.CANCEL)}>Valider cette affectation</PlainButton>
-      </div>
+      {user.role !== ROLES.REFERENT_DEPARTMENT && (
+        <div className="flex items-end justify-center pt-[67px]">
+          <BorderButton onClick={() => onChangeStep(GROUPSTEPS.MODIFICATION)} className="mr-[8px]">
+            Annuler
+          </BorderButton>
+          <PlainButton onClick={() => onChange(group, GROUPSTEPS.CANCEL)}>Valider cette affectation</PlainButton>
+        </div>
+      )}
     </GroupBox>
   );
 }